back back

+++ farm/species.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, redirect, render_template, request, url_for
+from flask import Blueprint, redirect, render_template, request, url_for, flash
 from farm.auth import login_required
 
 from flask_wtf import FlaskForm
@@ -25,15 +25,16 @@ class SpeciesForm(FlaskForm):
 def create():
     form = SpeciesForm()
     if request.method == 'POST':
+        idx = form.field.data
         sort_no = str((int(form.field.data) + 1) * 100 + int(form.sort_no.data))
         data = {
-            'field': form.field.data,
+            'field': idx,
             'family': form.family.data,
             'species': form.species.data,
             'sort_no': sort_no,
             }
         dao.create_species(data)
-        return redirect(url_for('index'))
+        return redirect(url_for('fields.docs', idx=idx))
 
     return render_template('forms/species.html', form=form)
 
@@ -48,12 +49,13 @@ def update(id):
     family = found['family']
     sort_no = found['sort_no']
     if request.method == 'POST':
-        found['field'] = form.field.data
+        idx = form.field.data
+        found['field'] = idx
         found['family'] = form.family.data
         found['species'] = form.species.data
         found['sort_no'] = form.sort_no.data
-        dao.update_species(id, found) 
-        return redirect(url_for('index'))
+        dao.update_species(id, found)
+        return redirect(url_for('fields.docs', idx=idx))
 
     return render_template('forms/species.html', 
                                 form=form,
@@ -67,5 +69,23 @@ def update(id):
 @bp.route('/delete/<id>')
 @login_required
 def delete(id):
-    dao.delete_species(id)
-    return redirect(url_for('index'))
+    count = dao.count_variety(id)
+    if not count:
+        dao.delete_species(id)
+        return redirect(url_for('index'))
+    else:
+        flash('Please remove the child documents first')
+        form = SpeciesForm()
+        found = dao.read_species(id)
+        field = found['field']
+        species = found['species']
+        family = found['family']
+        sort_no = found['sort_no']
+    return render_template('forms/species.html', 
+                                    form=form,
+                                    id=id,
+                                    field=field,
+                                    family=family, 
+                                    species=species,
+                                    sort_no=sort_no
+                                )
